feat(release prep script): make script more resilient (#4907)
* fix(release prep script): abort on local changes; fix version extraction; reset all checkboxes

* fix: better detection for of non-whitespace

* ref: better way to track local changes

* ref: don't take hidden files into account

* ref: update abort message

* feat: add heading depth when encountering heading after Tests section

+++ scripts/release_prep.sh
@@ -8,6 +8,15 @@ set -x
 # - PRs use test section LAST with heading "## Tests"
 # - ALL build and release PRs start with "build: "
 
+has_local_changes=$(git status --porcelain --untracked-files=no --ignored=no)
+if [[ ${has_local_changes} ]]; then
+  set +x
+  echo ==========
+  echo "ABORT: You have local modifications. Please stash or commit changes and run again."
+  echo ==========
+  exit 1
+fi
+
 git fetch --all --tags
 git reset --hard
 git pull
@@ -19,7 +28,7 @@ temp_release_branch=temp_${short_hash}
 
 git checkout -b ${temp_release_branch}
 
-release_version=$(npm --no-git-tag-version version minor | tail -n 1)
+release_version=$(npm --no-git-tag-version version minor | grep -E '^v\d')
 release_branch=release_${release_version}
 may_force_push=
 
@@ -67,8 +76,8 @@ echo "## Tests" >> ${pr_body_file_groupped}
 echo "" >> ${pr_body_file_groupped}
 grep -v -E -- '- [a-z]+\(deps(-dev)?\)' ${pr_body_file} | grep -v -E -- '- build: ' | while read line_item; do
   pr_id=$(echo ${line_item} | grep -o -E '\[`#\d+`\]' | grep -o -E '\d+')
-  tests=$(gh pr view ${pr_id} | awk 'f;/^#+ Tests?/{f=1}' | sed "s/\[X\]/[ ]/")
-  if [[ ${tests} =~ \S ]]; then
+  tests=$(gh pr view ${pr_id} | awk 'f;/^#+ Tests?/{f=1}' | sed -E "s/\[[Xx]\]/[ ]/" | sed -E "s/^(##+) /\1## /")
+  if [[ ${tests} =~ [^[:space:]] ]]; then
     echo ${line_item} | sed "s/^- /### /" >> ${pr_body_file_groupped}
     echo "${tests}" >> ${pr_body_file_groupped}
     echo "" >> ${pr_body_file_groupped}
